CLN: Move readers/writers to `readers`/`writers`
