+++ src/fmu/dataio/dataio.py
@@ -24,10 +24,7 @@ from ._logging import null_logger
 from ._metadata import generate_export_metadata
 from ._utils import (
     detect_inside_rms,  # dataio_examples,
-    export_file,
-    export_metadata_file,
     prettyprint_dict,
-    read_metadata_from_file,
     some_config_from_env,
 )
 from .aggregation import AggregatedData
@@ -36,6 +33,8 @@ from .datastructure._internal.internal import AllowedContent
 from .datastructure.configuration import global_configuration
 from .datastructure.meta import meta
 from .providers._fmu import FmuProvider, get_fmu_context_from_environment
+from .readers import read_metadata_from_file
+from .writers import export_file, export_metadata_file
 
 # DATAIO_EXAMPLES: Final = dataio_examples()
 INSIDE_RMS: Final = detect_inside_rms()
